src/sdk/python_sdk/pyproject.toml
[project]
name = "python-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "kafka-python>=2.2.15",
    "psutil>=7.0.0",
    "python-decouple>=3.8",
]
